Style and Quality Fixes
