+++ .gitignore
@@ -3,4 +3,5 @@
 vendor/
 phpunit.xml
 .php_cs.cache
+.phpunit.result.cache
