--- .phpunit.result.cache
@@ -1,1 +1,0 @@
-C:37:"PHPUnit\Runner\DefaultTestResultCache":399:{a:2:{s:7:"defects";a:2:{s:55:"Whisky\Test\SimpleExpressionsTest::testAssignExpression";i:4;s:56:"Whisky\Test\SimpleExpressionsTest::testForeachExpression";i:4;}s:5:"times";a:3:{s:55:"Whisky\Test\SimpleExpressionsTest::testAssignExpression";d:0.174;s:56:"Whisky\Test\SimpleExpressionsTest::testForeachExpression";d:0.104;s:61:"Whisky\Test\SimpleExpressionsTest::testFunctionCallExpression";d:0.123;}}}
