+++ src/Builder.php
@@ -6,5 +6,7 @@ namespace Whisky;
 
 interface Builder
 {
+    public function addExtension(Extension $extension): void;
+
     public function build(string $code): Script;
 }
