+++ src/Extension.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Whisky;
+
+interface Extension
+{
+    public function parse(string $code): void;
+
+    public function normalize(string $code): string;
+}
