+++ src/ParseError.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Whisky;
+
+
+class ParseError extends \Exception
+{
+
+}
